Support experimental Explore UI
The Explore UI is in beta for Grafana 5.3.1. I thought it would be
nice if we could support arbitrary PNG rendering for any explore query.

We accomplish this by basically looking for a rendered canvas (because
the Explore UI doesn't emit the same events as a rendered dashboard).

Then, since there is no solo mode for explore panels, we render the
graph element instead.