fix(go): propagate traceparent without respecting CORS
CORS headers ensure that we cannot always send `traceparent` via the
browser, according to the browser. However, we want to always send the
header regardless of CORS, such that the receiving service can decide
whether it wants to care about it.

As such, we should use the `Fetch` APIs, which let us arbitrarily change
the outgoing requests without any CORS restrictions.

Thanks to Josh for helping me figure this out!

Co-Authored-By: Josh Hunt <joshhunt@users.noreply.github.com>