fix(go): all actions should be observed
We currently do not use a wrapper for _all_ actions because we spawn our
own trace spans in our own code. However, this isn't all it does: it
also marks a metric, which we currently are missing on the most
important pieces.

As such, this renames the function so we don't misconstrue what it does,
and use it for _all_ actions (even our own!). This will improve our
metric outputs.