feat!: migrate to Go
BREAKING CHANGE: This migrates the entire service over to Go by default,
rather than the Node.js implementation. This means the entire service
has been rewritten (but with the same acceptance tests).

This comes with a whole lot more configurability, and enables us to
implement much stronger security measures in the service.